setHull speed improvement

+++ Vec2.h
@@ -253,7 +253,7 @@ private:
 
     static int orientation(const Vec2& p, const Vec2& q, const Vec2& r)
     {
-        int val = int((q.y - p.y) * (r.x - q.x) - (q.x - p.x) * (r.y - q.y));
+        float val = (q.y - p.y) * (r.x - q.x) - (q.x - p.x) * (r.y - q.y);
         if (val == 0) return 0;  // collinear
         return (val > 0) ? 1 : 2; // clock or counterclock wise
     }
